+++ libexec/ftpd/ftpd.c
@@ -3576,7 +3576,8 @@ logxfer(const char *command, off_t bytes, const char *file1, const char *file2,
 			    " %s", r2);
 		if (elapsed != NULL)
 			len += snprintf(buf + len, sizeof(buf) - len,
-			    " in %ld.%.03d seconds", elapsed->tv_sec,
+			    " in %lld.%.03d seconds",
+			    (long long int)elapsed->tv_sec,
 			    (int)(elapsed->tv_usec / 1000));
 		if (error != NULL)
 			len += snprintf(buf + len, sizeof(buf) - len,
@@ -3599,7 +3600,7 @@ logxfer(const char *command, off_t bytes, const char *file1, const char *file2,
 
 	time(&now);
 	len = snprintf(buf, sizeof(buf),
-	    "%.24s %ld %s " LLF " %s %c %s %c %c %s FTP 0 * %c\n",
+	    "%.24s %lld %s " LLF " %s %c %s %c %c %s FTP 0 * %c\n",
 
 /*
  * XXX: wu-ftpd puts ' (send)' or ' (recv)' in the syslog message, and removes
@@ -3607,7 +3608,7 @@ logxfer(const char *command, off_t bytes, const char *file1, const char *file2,
  *	given that syslog messages don't contain the full date.
  */
 	    ctime(&now),
-	    elapsed == NULL ? 0 : elapsed->tv_sec + (elapsed->tv_usec > 0),
+	    elapsed == NULL ? 0LL : elapsed->tv_sec + (elapsed->tv_usec > 0),
 	    remotehost,
 	    (LLT) bytes,
 	    r1,
@@ -3647,9 +3648,9 @@ logrusage(const struct rusage *rusage_before,
 
 	timersub(&rusage_after->ru_utime, &rusage_before->ru_utime, &usrtime);
 	timersub(&rusage_after->ru_stime, &rusage_before->ru_stime, &systime);
-	syslog(LOG_INFO, "%ld.%.03du %ld.%.03ds %ld+%ldio %ldpf+%ldw",
-	    usrtime.tv_sec, (int)(usrtime.tv_usec / 1000),
-	    systime.tv_sec, (int)(systime.tv_usec / 1000),
+	syslog(LOG_INFO, "%lld.%.03du %lld.%.03ds %ld+%ldio %ldpf+%ldw",
+	    (long long int)usrtime.tv_sec, (int)(usrtime.tv_usec / 1000),
+	    (long long int)systime.tv_sec, (int)(systime.tv_usec / 1000),
 	    rusage_after->ru_inblock - rusage_before->ru_inblock,
 	    rusage_after->ru_oublock - rusage_before->ru_oublock,
 	    rusage_after->ru_majflt - rusage_before->ru_majflt,
