+++ usr.sbin/makefs/makefs.h
@@ -220,8 +220,8 @@ extern	struct timespec	start_time;
 		struct timeval end, td;			\
 		gettimeofday(&end, NULL);		\
 		timersub(&end, &(x), &td);		\
-		printf("%s took %ld.%06ld seconds\n",	\
-		    (d), td.tv_sec, td.tv_usec);	\
+		printf("%s took %lld.%06ld seconds\n",	\
+		    (d), (long long int)td.tv_sec, td.tv_usec);	\
 	}
